Tower_Hanoi Problem with Detailed Explanation
This was an assignment in my Object-Oriented C++ course. Though it does not really involve classes, it taught me a lot about vectors and how to solve algorithms systematically. Drawing the towers on a paper helps a lot to visualize the steps and labeling them. I did not get it the first time I did it of course rather after doing it I went over it again leaving comments to make sure that I understand everything prior to the final exam of the class.